[REFACTOR] #17 - changed NewStorage name to Storage
