+++ EasyVel/Presentation/Scrap/Storage/ViewController/StorageViewController.swift
@@ -1,197 +1,244 @@
 //
 //  StorageViewController.swift
-//  VelogOnMobile
+//  EasyVel
 //
-//  Created by 홍준혁 on 2023/05/04.
+//  Created by 이성민 on 10/5/23.
 //
 
 import UIKit
 
 import RxSwift
 import RxCocoa
-
 import SnapKit
 
-final class StorageViewController: RxBaseViewController<StorageViewModel>, FolderViewControllerDelegate {
+final class StorageViewController: BaseViewController {
     
+    typealias Cell = StorageCollectionViewCell
+    typealias Header = StorageCollectionViewHeader
+    typealias DataSource = UICollectionViewDiffableDataSource<Section, StoragePost>
+    typealias Snapshot = NSDiffableDataSourceSnapshot<Section, StoragePost>
     
-    private let storageView = StorageView()
-    private var storagePosts: [StoragePost]?
-    private var storageTableViewDidScroll = false
-    var folderName: String?
-
-    override func render() {
+    enum Section {
+        case main
+    }
+    
+    // MARK: - Property
+    
+    private let viewModel: StorageViewModel
+    private let disposeBag = DisposeBag()
+    
+    private let storageView: StoragePostView
+    private var storageDataSource: DataSource!
+    private var storageSnapshot: Snapshot!
+    
+    private var changeFolderNameAlert: UIView?
+    private var deleteFolderBottomSheet: UIView?
+    
+    // MARK: - Life Cycle
+    
+    init(
+        view: StoragePostView,
+        viewModel: StorageViewModel
+    ) {
+        self.storageView = view
+        self.viewModel = viewModel
+        super.init(nibName: nil, bundle: nil)
+    }
+    
+    @available(*, unavailable)
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    override func loadView() {
+        super.loadView()
         self.view = storageView
     }
-  
-    override func bind(viewModel: StorageViewModel) {
-        super.bind(viewModel: viewModel)
-        bindOutput(viewModel)
-        setDelegate()
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
         
-        storageView.storageHeadView.deleteFolderButton.rx.tap
-            .subscribe(onNext: { [weak self] in
-                self?.presentDeleteFolderActionSheet()
-            })
-            .disposed(by: disposeBag)
+        self.setDataSource()
+        self.setSnapshot()
+        self.bind()
+        self.bindViewModel()
+        self.bindNavigation()
+    }
+    
+    // MARK: - Setting
+    
+    private func bind() {
+        self.storageView.backButton.rx.tap
+            .subscribe(onNext: {
+                self.navigationController?.popViewController(animated: true)
+            }).disposed(by: self.disposeBag)
         
-        storageView.storageHeadView.changeFolderNameButton.rx.tap
-            .subscribe(onNext: { [weak self] in
-                
-                let vc = FolderAlertViewController(type: .change, folderName: self?.folderName ?? "")
-                vc.modalPresentationStyle = .overFullScreen
-                vc.modalTransitionStyle = .crossDissolve
-                vc.delegate = self
-                self?.present(vc, animated: false)
-            })
-            .disposed(by: disposeBag)
+        self.storageView.dismissTrigger
+            .drive(with: self) { owner, _ in
+                self.storageView.dismissDeleteFolderBottomSheet()
+            }.disposed(by: self.disposeBag)
+    }
+    
+    private func bindViewModel() {
+        let viewWillAppear = self.rx.viewWillAppear
+            .asDriver()
+        let deleteFolderTrigger = self.storageView.deleteFolderTrigger
+            .asDriver()
         
-        storageView.listTableView.rx.contentOffset
-            .subscribe(onNext: { [weak self] contentOffset in
-                let scrollY = contentOffset.y
-                let isAllScrapFolder = self?.viewModel?.folderName == TextLiterals.allPostsScrapFolderText ? true : false
-                if scrollY > 5 && self?.storageTableViewDidScroll == false {
-                    self?.storageView.storageTableViewStartScroll()
-                    self?.storageTableViewDidScroll.toggle()
-                } else if scrollY < 2 && self?.storageTableViewDidScroll == true {
-                    self?.storageView.storageTableViewEndScroll(
-                        isAllpostFolder: isAllScrapFolder
-                    )
-                    self?.storageTableViewDidScroll.toggle()
-                }
-            })
-            .disposed(by: disposeBag)
+        let input = StorageViewModel.Input(viewWillAppear, deleteFolderTrigger)
+        let output = self.viewModel.transform(input)
         
-        storageView.storageHeadView.viewPopButton.rx.tap
-            .subscribe(onNext: { [weak self] in
-                self?.navigationController?.popViewController(animated: true)
-            })
-            .disposed(by: disposeBag)
-    }
-    
-    private func bindOutput(_ viewModel: StorageViewModel) {
-        viewModel.storagePostsOutput
-            .asDriver(onErrorJustReturn: [])
-            .drive(onNext: { [weak self] post in
-                self?.storagePosts = post
-                self?.storageView.listTableView.reloadData()
-            })
-            .disposed(by: disposeBag)
+        output.storagePosts
+            .drive(with: self) { owner, posts in
+                if posts.isEmpty { self.storageView.showEmptyView() }
+                else { owner.loadSnapshot(with: posts) }
+            }.disposed(by: self.disposeBag)
         
-        viewModel.isPostsEmptyOutput
-            .asDriver(onErrorJustReturn: Bool())
-            .drive(onNext: { [weak self] isPostsEmpty in
-                if isPostsEmpty {
-                    self?.storageView.storageViewExceptionView.isHidden = false
-                } else {
-                    self?.storageView.storageViewExceptionView.isHidden = true
-                }
-            })
-            .disposed(by: disposeBag)
+        output.folderDeleted
+            .drive(with: self) { owner, _ in
+                owner.storageView.dismissDeleteFolderBottomSheet()
+                owner.showToastOnParentViewController(type: .folderDeleted)
+            }.disposed(by: self.disposeBag)
+    }
+    
+    private func bindCell(_ cell: Cell) {
+        let editStorageStatusTrigger = cell.editPostStatusTrigger
+            .asDriver()
         
-        viewModel.folderNameOutput
-            .asDriver(onErrorJustReturn: String())
-            .drive(onNext: { [weak self] folderName in
-                self?.folderName = folderName
-                if folderName == TextLiterals.allPostsScrapFolderText {
-                    self?.storageView.storageHeadView.deleteFolderButton.isHidden = true
-                    self?.storageView.storageHeadView.changeFolderNameButton.isHidden = true
-                } else {
-                    self?.storageView.storageHeadView.deleteFolderButton.isHidden = false
-                    self?.storageView.storageHeadView.changeFolderNameButton.isHidden = false
-                }
-            })
-            .disposed(by: disposeBag)
-
+        let input = StorageViewModel.CellInput(editStorageStatusTrigger)
+        let output = self.viewModel.transformCell(input)
+        
+        output.newPosts.drive(
+            with: self,
+            onNext: { owner, storagePosts in
+                owner.loadSnapshot(with: storagePosts)
+                Toast.show(
+                    toastText: TextLiterals.postDescrappedToast,
+                    toastBackgroundColor: .gray500,
+                    controller: self
+                )
+            }).disposed(by: cell.disposeBag)
     }
     
-    func folderVCDismiss(newFolderName:String) {
-        self.folderName = newFolderName
-        self.storageView.storageHeadView.titleLabel.text = newFolderName
+    private func bindHeader(_ header: Header) {
+        let changeNameTrigger = header.changeNameButtonTrigger
+            .asDriver()
+        
+        let input = StorageViewModel.HeaderInput(changeNameTrigger)
+        let output = self.viewModel.transformHeader(input)
+        
+        output.showChangeNameAlert
+            .drive(with: self) { owner, folderName in
+                owner.showChageFolderNameAlert(folderName: folderName)
+            }.disposed(by: self.disposeBag)
     }
     
-    
-    func setStorageHeadView(
-        headTitle: String
-    ) {
-        storageView.storageHeadView.titleLabel.text = headTitle
-    }
-    
-    private func setDelegate() {
-        storageView.listTableView.dataSource = self
-        storageView.listTableView.delegate = self
-    }
-    
-    private func presentDeleteFolderActionSheet() {
-        let actionSheetController = UIAlertController(
-            title: TextLiterals.deleteFolderActionSheetTitle,
-            message: TextLiterals.deleteFolderActionSheetMessage,
-            preferredStyle: .actionSheet
-        )
-        let actionDefault = UIAlertAction(
-            title: TextLiterals.deleteFolderActionSheetOkActionText,
-            style: .destructive,
-            handler: { [weak self] _ in
-                self?.viewModel?.deleteFolderButtonDidTap.accept(true)
-                self?.navigationController?.popViewController(animated: true)
-            })
-        let actionCancel = UIAlertAction(
-            title: TextLiterals.deleteFolderActionSheetCancelActionText,
-            style: .cancel
-        )
-        actionSheetController.addAction(actionDefault)
-        actionSheetController.addAction(actionCancel)
-        self.present(actionSheetController, animated: true)
-    }
-
-}
-
-extension StorageViewController: UITableViewDataSource {
-    func numberOfSections(in tableView: UITableView) -> Int {
-        return storagePosts?.count ?? 0
-    }
-    
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return 1
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: StorageTableViewCell.identifier, for: indexPath) as? StorageTableViewCell else {
-            return StorageTableViewCell()
-        }
-        cell.selectionStyle = .none
-        cell.deleteButtonTappedClosure = { [weak self] url in
-            self?.viewModel?.deletePostButtonDidTap.accept(url)
-        }
-        let index = indexPath.section
-        if let data = storagePosts?[index] {
-            cell.binding(model: data)
-            return cell
-        }
-        return cell
+    private func bindNavigation() {
+        self.storageView.collectionView.rx.itemSelected
+            .bind(with: self, onNext: { owner, indexPath in
+                let collectionView = owner.storageView.collectionView
+                guard let postCell = collectionView.cellForItem(at: indexPath) as? Cell,
+                      let url = postCell.post?.url
+                else { return }
+                owner.pushToWebView(url)
+            }).disposed(by: self.disposeBag)
     }
 }
 
-extension StorageViewController: UITableViewDelegate {
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        let textNum = storagePosts?[indexPath.section].summary?.count ?? 0
-        if storagePosts?[indexPath.section].img ?? String() == TextLiterals.noneText {
-            switch textNum {
-            case 0...50: return SizeLiterals.postCellSmall
-            case 51...80: return SizeLiterals.postCellMedium
-            case 81...100: return SizeLiterals.postCellLarge
-            default: return SizeLiterals.postCellLarge
-            }
-        } else {
-            return SizeLiterals.postCellXLarge
-        }
-    }
-    
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        let selectedCell = tableView.cellForRow(at: indexPath) as? StorageTableViewCell
-        guard let url = selectedCell?.url else { return }
+// MARK: - Custom Method
+
+extension StorageViewController {
+    private func pushToWebView(_ url: String) {
         let webViewModel = WebViewModel(url: url, service: DefaultFollowService.shared)
         let webViewController = WebViewController(viewModel: webViewModel)
-        navigationController?.pushViewController(webViewController, animated: true)
+        self.navigationController?.pushViewController(webViewController, animated: true)
+    }
+    
+    private func showToastOnParentViewController(type: ToastType) {
+        guard let parent = self.parent else { return }
+        self.navigationController?.popViewController(animated: true)
+        self.showToast(of: type, on: parent)
+    }
+}
+
+// MARK: - DataSource
+
+extension StorageViewController {
+    
+    private func setDataSource() {
+        self.configureDataSource()
+        self.configureDataSourceHeader()
+    }
+    
+    private func configureDataSource() {
+        let cellRegsitration = UICollectionView.CellRegistration<Cell, StoragePost> { [weak self] cell, _, post in
+            cell.loadPost(post)
+            self?.bindCell(cell)
+        }
+        
+        self.storageDataSource = DataSource(collectionView: self.storageView.collectionView) { collectionView, indexPath, post in
+            return collectionView.dequeueConfiguredReusableCell(
+                using: cellRegsitration,
+                for: indexPath,
+                item: post
+            )
+        }
+    }
+    
+    private func configureDataSourceHeader() {
+        let headerRegistration = UICollectionView.SupplementaryRegistration<Header>(
+            elementKind: UICollectionView.elementKindSectionHeader
+        ) { [weak self] header, _, _ in
+            self?.bindHeader(header)
+        }
+        
+        self.storageDataSource.supplementaryViewProvider = { _, _, indexPath in
+            return self.storageView.collectionView.dequeueConfiguredReusableSupplementary(
+                using: headerRegistration,
+                for: indexPath
+            )
+        }
+    }
+}
+
+// MARK: - Snapshot
+
+extension StorageViewController {
+    
+    private func setSnapshot() {
+        self.storageSnapshot = Snapshot()
+        self.storageSnapshot.appendSections([.main])
+        self.storageDataSource.apply(self.storageSnapshot)
+    }
+    
+    private func loadSnapshot(
+        with storagePosts: [StoragePost]
+    ) {
+        let previousPosts = self.storageSnapshot.itemIdentifiers(inSection: .main)
+        self.storageSnapshot.deleteItems(previousPosts)
+        self.storageSnapshot.appendItems(storagePosts, toSection: .main)
+        self.storageDataSource.apply(self.storageSnapshot, animatingDifferences: true)
+    }
+}
+
+// MARK: - Alerts
+
+extension StorageViewController {
+    
+    private func showChageFolderNameAlert(folderName: String) {
+        let vc = FolderAlertViewController(type: .change, folderName: folderName)
+        vc.modalPresentationStyle = .overFullScreen
+        vc.modalTransitionStyle = .crossDissolve
+        vc.delegate = self
+        self.present(vc, animated: false)
+    }
+    
+}
+
+// MARK: - FolderViewControllerDelegate
+
+extension StorageViewController: FolderViewControllerDelegate {
+    func folderVCDismiss(newFolderName: String) {
+        self.storageView.updateTitle(to: newFolderName)
     }
 }
