+++ EasyVel/Presentation/Scrap/Storage/ViewModel/StorageViewModel.swift
@@ -1,152 +1,142 @@
 //
-//  StorageViewModel.swift
-//  VelogOnMobile
+//  NewStorageViewModel.swift
+//  EasyVel
 //
-//  Created by 홍준혁 on 2023/05/05.
+//  Created by 이성민 on 10/5/23.
 //
 
 import Foundation
 
-import RealmSwift
-import RxRelay
 import RxSwift
+import RxCocoa
+import RxRelay
 
 final class StorageViewModel: BaseViewModel {
-
+    
+    // MARK: - Properties
+    
     let realm = RealmService()
-    var folderName: String?
+    
+    private var folderName: String
+    private var unScrappedPostURLs: [String] = []
+    
+    // MARK: - Input & Output
+    
+    struct Input {
+        let fetchPostTrigger: Driver<Void>
+        let deleteFolderTrigger: Driver<Void>
+        
+        init(_ fetchPostTrigger: Driver<Void>,
+             _ deleteFolderTrigger: Driver<Void>) {
+            self.fetchPostTrigger = fetchPostTrigger
+            self.deleteFolderTrigger = deleteFolderTrigger
+        }
+    }
+    
+    struct Output {
+        let storagePosts: Driver<[StoragePost]>
+        let folderDeleted: Driver<Void>
+        
+        init(_ storagePosts: Driver<[StoragePost]>,
+             _ folderDeleted: Driver<Void>) {
+            self.storagePosts = storagePosts
+            self.folderDeleted = folderDeleted
+        }
+    }
+    
+    struct CellInput {
+        let editPostStatusTrigger: Driver<String>
+        
+        init(_ editPostStatusTrigger: Driver<String>) {
+            self.editPostStatusTrigger = editPostStatusTrigger
+        }
+    }
+    
+    struct CellOutput {
+        let newPosts: Driver<[StoragePost]>
+        
+        init(_ newPosts: Driver<[StoragePost]>) {
+            self.newPosts = newPosts
+        }
+    }
+    
+    struct HeaderInput {
+        let changeNameTrigger: Driver<Void>
+        
+        init(_ changeNameTrigger: Driver<Void>) {
+            self.changeNameTrigger = changeNameTrigger
+        }
+    }
+    
+    struct HeaderOutput {
+        let showChangeNameAlert: Driver<String>
+        
+        init(_ showChangeNameAlert: Driver<String>) {
+            self.showChangeNameAlert = showChangeNameAlert
+        }
+    }
+    
+    // MARK: - Initialize
+    
+    init(folderName: String) {
+        self.folderName = folderName
+    }
+    
+    // MARK: - Transform
+    
+    func transform(_ input: Input) -> Output {
+        let storagePosts = input.fetchPostTrigger
+            .map { [weak self] _ -> [StoragePost] in
+                guard let self = self else { return [] }
+                return self.getRealmStoragePosts(from: self.folderName)
+            }
+        
+        let folderDeleted = input.deleteFolderTrigger
+            .map { [weak self] in
+                guard let self = self else { return }
+                self.deleteStorageFolder(of: self.folderName)
+            }
+        
+        return Output(storagePosts, folderDeleted)
+    }
+    
+    func transformCell(_ input: CellInput) -> CellOutput {
+        let newStoragePosts = input.editPostStatusTrigger
+            .map { [weak self] url -> [StoragePost] in
+                guard let self = self else { return [] }
+                self.realm.deletePost(url: url)
+                return self.getRealmStoragePosts(from: self.folderName)
+            }
+        
+        return CellOutput(newStoragePosts)
+    }
+    
+    func transformHeader(_ input: HeaderInput) -> HeaderOutput {
+        let showChangeNameAlert = input.changeNameTrigger
+            .map { [weak self] in
+                guard let self = self else { return String() }
+                return self.folderName
+            }
+        
+        return HeaderOutput(showChangeNameAlert)
+    }
+}
 
-    // MARK: - Input
-    
-    var deletePostButtonDidTap = PublishRelay<String>()
-    var deleteFolderButtonDidTap = PublishRelay<Bool>()
-    var changeFolderButtonDidTap = PublishRelay<([StoragePost], String)>()
-    var yesDidTap = PublishRelay<Bool>()
-    
-    // MARK: - Output
-    
-    var storagePostsOutput = PublishRelay<[StoragePost]>()
-    var isPostsEmptyOutput = PublishRelay<Bool>()
-    var folderNameOutput = PublishRelay<String>()
-    var newFolderNameIsUniqueOutput = PublishRelay<(String, Bool)>()
-    var newFolderOutput = PublishRelay<String>()
-    
-    override init() {
-        super.init()
-        makeOutput()
-    }
-    
-    private func makeOutput() {
-        viewWillAppear
-            .subscribe(onNext: { [weak self] in
-                let realmData = self?.getFolderPostInRealm(
-                    folderName: self?.folderName ?? ""
-                )
-                self?.storagePostsOutput.accept(realmData ?? [StoragePost]())
-                let isEmpty = self?.checkStorageEmpty(
-                    storage: realmData ?? [StoragePost]()
-                )
-                self?.isPostsEmptyOutput.accept(isEmpty ?? Bool())
-                self?.folderNameOutput.accept(self?.folderName ?? String())
-            })
-            .disposed(by: disposeBag)
-        
-        deletePostButtonDidTap
-            .subscribe(onNext: { [weak self] url in
-                self?.realm.deletePost(url: url)
-                let folderRealmData = self?.getFolderPostInRealm(
-                    folderName: self?.folderName ?? ""
-                )
-                let isfolderEmpty = self?.checkStorageEmpty(
-                    storage: folderRealmData ?? [StoragePost]()
-                )
-                self?.isPostsEmptyOutput.accept(isfolderEmpty ?? Bool())
-                self?.storagePostsOutput.accept(folderRealmData ?? [StoragePost]())
-            })
-            .disposed(by: disposeBag)
-        
-        deleteFolderButtonDidTap
-            .subscribe(onNext: { [weak self] didFolderDelete in
-                if didFolderDelete {
-                    self?.realm.deleteFolder(folderName: self?.folderName ?? String())
-                }
-            })
-            .disposed(by: disposeBag)
-        
-        changeFolderButtonDidTap
-            .subscribe(onNext: { [weak self] storagePosts, changeFolderName in
-                guard let oldFolderName = self?.folderName else { return }
-                guard let isNewFolderNameUnique = self?.checkIsUniqueFolderName(newFolderName: changeFolderName) else {
-                    return
-                }
-                if changeFolderName == "" {
-                    return
-                }
-                if isNewFolderNameUnique {
-                    self?.newFolderNameIsUniqueOutput.accept(
-                        (changeFolderName, true)
-                    )
-                    self?.realm.changeFolderNameInStorage(
-                        input: storagePosts,
-                        oldFolderName: oldFolderName,
-                        newFolderName: changeFolderName
-                    )
-                    self?.folderName = changeFolderName
-                } else {
-                    self?.newFolderNameIsUniqueOutput.accept(
-                        (changeFolderName, false)
-                    )
-                }
-            })
-            .disposed(by: disposeBag)
-        
-        yesDidTap
-            .subscribe(onNext: { [weak self] _ in
-                let realmData = self?.getFolderPostInRealm(
-                    folderName: self?.folderName ?? ""
-                )
-                self?.storagePostsOutput.accept(realmData ?? [StoragePost]())
-                let isEmpty = self?.checkStorageEmpty(
-                    storage: realmData ?? [StoragePost]()
-                )
-                self?.isPostsEmptyOutput.accept(isEmpty ?? Bool())
-                self?.folderNameOutput.accept(self?.folderName ?? String())
-            })
-            .disposed(by: disposeBag)
-       
-    }
-    
-    // MARK: - func
-    
-    private func checkIsUniqueFolderName(
-        newFolderName: String
-    ) -> Bool {
-        let isUniqueFolderName = self.realm.checkUniqueFolderName(
-            newFolderName: newFolderName
-        )
-        return isUniqueFolderName
-    }
-    
-    private func getPostInRealm() -> [StoragePost] {
-        let realmPostData = realm.getPosts()
-        let posts: [StoragePost] = realm.convertToStoragePost(input: realmPostData)
-        let reversePosts = realm.reversePosts(input: posts)
-        return reversePosts
-    }
-    
-    private func getFolderPostInRealm(
-        folderName: String
-    ) -> [StoragePost] {
-        let realmPostData = realm.getFolderPosts(folderName: folderName)
-        let posts: [StoragePost] = realm.convertToStoragePost(input: realmPostData)
-        let reversePosts = realm.reversePosts(input: posts)
-        return reversePosts
-    }
+// MARK: - Realm Functions
 
-    private func checkStorageEmpty(
-        storage: [StoragePost]
-    ) -> Bool {
-        if storage.count == 0 { return true }
-        else { return false }
+extension StorageViewModel {
+    
+    private func getRealmStoragePosts(from name: String) -> [StoragePost] {
+        let realmPosts = self.realm.getFolderPosts(folderName: name)
+        let storagePosts = self.realm.convertToStoragePost(input: realmPosts)
+        return storagePosts.reversed()
+    }
+    
+    private func isFolderNameUnique(_ newName: String) -> Bool {
+        return self.realm.checkUniqueFolderName(newFolderName: newName)
+    }
+    
+    private func deleteStorageFolder(of name: String) {
+        return self.realm.deleteFolder(folderName: name)
     }
 }
